Separated Parse-Json.ts into individual files for AbilityScore, Conditions, Skill, and Spell
